object_detection_rnn/requirements.txt
python
numpy
tqdm
pytorch
torchvision
matplotlib
